Account for https://github.com/typed-ember/glint/issues/697
Update README

README

Add vitest

oof
