+++ package.json
@@ -57,7 +57,8 @@
     "prettier": "^3.2.4",
     "publint": "^0.2.7",
     "release-plan": "^0.7.0",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "vitest": "^1.2.2"
   },
   "publishConfig": {
     "access": "public",
