+++ src/fixes/ember.js
@@ -1,6 +1,4 @@
-import jscodeshift from 'jscodeshift';
-
-const j = jscodeshift.withParser('ts');
+import { fixReferences } from './typescript.js';
 
 const defaultFind = 'ember';
 
@@ -9,17 +7,7 @@ const defaultFind = 'ember';
  * @param {{ types?: string }} [ options ]
  */
 export function fixEmberReferences(contents, options = {}) {
-  const root = j(contents);
-  const find = `/ <reference types="${options.types || defaultFind}`;
-
-  const fixed = root
-    // @ts-expect-error
-    .find(j.Comment)
-    // @ts-expect-error
-    .filter((path) => path.value.value.startsWith(find))
-    // @ts-expect-error
-    .forEach((path) => j(path).remove())
-    .toSource();
-
-  return fixed;
+  return fixReferences(contents, {
+    types: options.types || defaultFind,
+  });
 }
